+++ esap/courses/seguridadInformatica/unidad2.php
@@ -1,4 +1,4 @@
-<?php require_once '../../../config.php'; ?>
+<?php require_once '../../config.php'; ?>
 <div class="col-sm-12">
     contenido de la unidad 2
 </div>
